requirements.txt
streamlit
langchain-community
groq
PyPDF2
unstructured